Haproxy mksite/mkconfig simplification (#28)
* Migrate haproxy config file housekeeping from mkconfig.bash to entrypoint

* Move docker-compose specific envar defs into docker-compose.yml file

fix metrics_path for custom web.external-url

Fix up shim mode to match recent changes in normal mode; fix regression in migration 3

Fix shell metacharacter parsing issue

safer .env file handling in scripts; not all systems have base32, use tr instead; many double-quote warnings

Shim mode no longer includes prometheus

+++ contrib/docker-compose/standalone/init-letsencrypt.bash
@@ -26,17 +26,24 @@
 
 HERE=$(cd "$(dirname "$0")"; pwd)
 set -eu
-[ -e "$HERE/.env" ]
-. "$HERE/.env"
+[ -f "$HERE/.env" ] || { echo "Environment file not found; you must run ./mksite.bash first"; exit 1; }
 
 if ! [ -x "$(command -v docker-compose)" ]; then
   echo 'Error: docker-compose is not installed.' >&2
   exit 1
 fi
 
+FQDN=$(awk -F= '/^FQDN=/ {print $2}' < "$HERE/.env" | tail -1)
+ALIAS_FQDNS=$(awk -F= '/^ALIAS_FQDNS=/ {print $2; exit}' < "$HERE/.env" | tail -1)
+
+# Strip enclosing quotes, as docker-compose<1.29 does not parse shell metachars in .env
+# See https://github.com/docker/compose/issues/8388
+ALIAS_FQDNS="${ALIAS_FQDNS%\"}"
+ALIAS_FQDNS="${ALIAS_FQDNS#\"}"
+
 domains=($FQDN $ALIAS_FQDNS)
 rsa_key_size=4096
-email="$EMAIL" # Adding a valid address is strongly recommended
+email=$(awk -F= '/^EMAIL=/ {print $2}' < "$HERE/.env" | tail -1)
 
 echo "### Downloading recommended TLS parameters ..."
 docker-compose run --rm --entrypoint "/bin/sh -c \"\
